Update Cadwyn migrations with v3-1-test changes
After cherry-picking, the versions needs to change on one of the Cadwyn migrations.
